backend/requirements.txt
sentence-transformers==2.2.2
pandas==2.0.1
numpy==1.23.4
PyPDF2==3.0.1
python-dotenv==1.0.0
qdrant-client==1.1.6